requirements.txt
numpy==1.17.4
opencv_python==4.5.5.64
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
